requirements.txt
apache-airflow-providers-snowflake==5.5.0
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.1
